Extract tweet item to dedicated component
